+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Added
+
+- Rovers are now recognized by the ArduPilot autopilot driver in the MAVLink
+  extension and the proper modes are displayed on the Live UI, thanks to
+  [@r0n4n](https://github.com/r0n4n) in PR
+  [#12](https://github.com/skybrush-io/skybrush-server/pull/12/files)
+
 ## [2.39.2] - 2025-11-25
 
 ### Fixed